# chap-money/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Chap Money checkout API — FedaPay direct mobile money + card redirect
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MobilePaymentInput {
  /** Amount in XOF (integer, no decimals) */
  amount: number;
  /** Payment description */
  description: string;
  /** Customer email — overridden server-side, can be omitted */
  customerEmail?: string;
  customerFirstname?: string;
  customerLastname?: string;
  /** Phone number without country prefix */
  phoneNumber: string;
  /** Two-letter country code lowercase */
  country: string;
  /** FedaPay operator mode code (mtn, moov, sbin, mtn_ci, free_sn, moov_tg, moov_bf, airtel_ne, mtn_open_gn) */
  operator: string;
}

export interface CardPaymentInput {
  /** Amount in XOF */
  amount: number;
  description: string;
  /** Customer fields — overridden server-side, can be omitted */
  customerEmail?: string;
  customerFirstname?: string;
  customerLastname?: string;
}

export interface CheckoutInitiated {
  /** FedaPay transaction ID */
  transactionId: number;
  /** Initial payment status (pending) */
  status: string;
  /** User-friendly confirmation message */
  message: string;
}

export interface CheckoutCard {
  transactionId: number;
  /** FedaPay hosted checkout URL for card payment */
  paymentUrl: string;
}

export interface PaymentStatus {
  transactionId: string;
  /** pending | approved | declined | cancelled | transferred */
  status: string;
}

export interface ErrorResponse {
  error: string;
}