# chap-money/lib/api-zod/src/generated/types/cardPaymentInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Chap Money checkout API — FedaPay direct mobile money + card redirect
 * OpenAPI spec version: 0.1.0
 */

export interface CardPaymentInput {
  /** Amount in XOF */
  amount: number;
  description: string;
  /** Overridden server-side */
  customerEmail?: string;
  customerFirstname?: string;
  customerLastname?: string;
}