# chap-money/lib/api-zod/src/generated/types/checkoutCard.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Chap Money checkout API — FedaPay direct mobile money + card redirect
 * OpenAPI spec version: 0.1.0
 */

export interface CheckoutCard {
  transactionId: number;
  /** FedaPay hosted checkout URL for card payment */
  paymentUrl: string;
}