# chap-money/lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Chap Money checkout API — FedaPay direct mobile money + card redirect
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
}