# chap-money/lib/api-zod/src/generated/types/mobilePaymentInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Chap Money checkout API — FedaPay direct mobile money + card redirect
 * OpenAPI spec version: 0.1.0
 */

export interface MobilePaymentInput {
  /** Amount in XOF (integer, no decimals) */
  amount: number;
  /** Payment description */
  description: string;
  /** Customer email — overridden server-side */
  customerEmail?: string;
  customerFirstname?: string;
  customerLastname?: string;
  /** Phone number without country prefix */
  phoneNumber: string;
  /** Two-letter country code lowercase */
  country: string;
  /** FedaPay operator mode code (mtn, moov, sbin, mtn_ci, free_sn, moov_tg, moov_bf, airtel_ne, mtn_open_gn) */
  operator: string;
}